pyproject.toml
[tool.poetry]
name = "loftstore"
version = "0.1.0"
description = ""
authors = ["Deni <denibatyzhev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.6.0"
attrs = "22.2.0"
autopep8 = "2.0.1"
build = "0.10.0"
cachecontrol = "0.12.11"
certifi = "2022.12.7"
charset-normalizer = "3.1.0"
cleo = "2.0.1"
colorama = "0.4.6"
coreapi = "2.3.3"
coreschema = "0.0.4"
crashtest = "0.4.1"
distlib = "0.3.6"
django = "4.1.7"
django-autoslug = "1.9.8"
django-extra-fields = "3.0.2"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
drf-spectacular = "0.26.0"
drf-writable-nested = "0.7.0"
drf-yasg = "1.21.5"
dulwich = "0.21.3"
filelock = "3.9.0"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "6.0.0"
inflection = "0.5.1"
installer = "0.6.0"
itypes = "1.2.0"
jaraco-classes = "3.2.3"
jinja2 = "3.1.2"
jsonschema = "4.17.3"
keyring = "23.13.1"
lockfile = "0.12.2"
markupsafe = "2.1.2"
more-itertools = "9.1.0"
msgpack = "1.0.4"
packaging = "23.0"
pexpect = "4.8.0"
pillow = "9.4.0"
pkginfo = "1.9.6"
platformdirs = "2.6.2"
poetry = "1.4.0"
poetry-core = "1.5.1"
poetry-plugin-export = "1.3.0"
psycopg2 = "2.9.5"
ptyprocess = "0.7.0"
pycodestyle = "2.10.0"
pyjwt = "2.6.0"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
python-slugify = "8.0.1"
pytz = "2022.7.1"
pywin32-ctypes = "0.2.0"
pyyaml = "6.0"
rapidfuzz = "2.13.7"
requests = "2.28.2"
requests-toolbelt = "0.10.1"
ruamel-yaml = "0.17.21"
shellingham = "1.5.0.post1"
six = "1.16.0"
sqlparse = "0.4.3"
text-unidecode = "1.3"
tomlkit = "0.11.6"
trove-classifiers = "2023.2.20"
tzdata = "2022.7"
uritemplate = "4.1.1"
urllib3 = "1.26.14"
virtualenv = "20.20.0"
webencodings = "0.5.1"
zipp = "3.15.0"
django-rest-passwordreset = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
